+++ src/platforms/ucp-mobile/setup/experiments/ucp-mobile-experiments.setup.ts
@@ -6,18 +6,12 @@ const OPTIMIZELY_PERFORMANCE_ADS_FOR_FANATICAL = {
 	EXPERIMENT_VARIANT: 'mobile_performance_ads_variant',
 };
 
-const OPTIMIZELY_OPEN_WEB_REACTIONS = {
-	EXPERIMENT_ENABLED: 'mobile_open_web',
-	EXPERIMENT_VARIANT: 'mobile_open_web_variant',
-};
-
 @Injectable()
 export class UcpMobileExperimentsSetup implements DiProcess {
 	constructor(private optimizely: Optimizely) {}
 
 	execute(): void {
 		this.configurePerformanceAdsExperiment();
-		this.configureOpenWebReactionsExperiment();
 	}
 
 	private configurePerformanceAdsExperiment(): void {
@@ -25,14 +19,6 @@ export class UcpMobileExperimentsSetup implements DiProcess {
 
 		if (variant) {
 			this.optimizely.addVariantToTargeting(OPTIMIZELY_PERFORMANCE_ADS_FOR_FANATICAL, variant);
-		}
-	}
-
-	private configureOpenWebReactionsExperiment(): void {
-		const variant = this.optimizely.getVariant(OPTIMIZELY_OPEN_WEB_REACTIONS);
-
-		if (variant) {
-			this.optimizely.addVariantToTargeting(OPTIMIZELY_OPEN_WEB_REACTIONS, variant);
 		}
 	}
 }
